fix: WhatsApp config — server auth guard, PhoneInput, polling stale closure (#420)
1. Split into server component (auth guard + redirect) + client component
2. Replace raw <Input> with <PhoneInput> (country selector + validation)
3. Fix polling stale closure using useRef for connectionStatus/qrCode
4. Replace hardcoded text-gray-* with semantic theme classes (dark mode)
